refactor(structure): bucket feature components by domain + scaffold architecture dirs
Phase 1 of the directory restructure (components only, no logic changes).

Moves (git mv, history preserved):
- Group 22 flat components under src/components/sections/features/ into
  domain buckets: investment/ tfff/ forest-cover/ endorsement/ partners/
  about/ shared/.
- Rename typo file FlowChartArrorw.tsx -> FlowChartArrow.tsx.
- Update all importers; fix forest-cover relative chart imports
  (../charts -> ../../charts) after the extra nesting level.

Scaffolding (empty + README, for the upcoming architecture phase):
- src/domain, src/content, src/lib, src/data
- components/maps/base, charts/primitives, charts/transforms
- READMEs per dir describe purpose, current contents, and what should
  migrate/delete in later phases (candidates C01-C08).

Also adds the docs/refactor/ plan (AI docs, directory, UI, architecture).

Deferred to later phases: static data moves, utils/->lib/, store
consolidation, all logic changes. Verified: tsc --noEmit clean,
next build green (31/31 pages), lint warnings pre-existing only.
