+++ src/components/maps/base/README.md
@@ -1,0 +1,3 @@
+# maps/base/
+
+Scaffold for the architecture phase (candidate C03 — `<BaseMap>`). Will own the shared MapLibre boot, responsive zoom, geofence clamp, attribution, and typed `onFeatureClick`, currently duplicated across `world/` and `country/`. Empty for now. See `docs/refactor/04-architecture.md`.
