+++ src/components/sections/charts/primitives/README.md
@@ -1,0 +1,3 @@
+# charts/primitives/
+
+Scaffold for the architecture phase (candidate C04). Will hold thin presentational chart primitives (`DonutChart`, `ChartTooltip`) wrapping recharts defaults, replacing per-chart duplication. Empty for now. See `docs/refactor/04-architecture.md`.
