+++ src/components/sections/charts/transforms/README.md
@@ -1,0 +1,3 @@
+# charts/transforms/
+
+Scaffold for the architecture phase (candidate C04). Will hold pure `(records, dataset) → ChartData` functions (no React) extracted from the inline `useMemo` blocks in the chart components. This becomes the unit-test surface. Empty for now. See `docs/refactor/04-architecture.md`.
