+++ src/components/sections/features/about/README.md
@@ -1,0 +1,9 @@
+# features/about/
+
+**Purpose:** Page-section components for the "About TFFF Watch" content.
+
+**Here now:** `AboutTFFFWatch`, `WhatsTFFFWatch`.
+
+**Note:** `WhatsTFFFWatch` has no live importer (commented out in `InfoGrid`) — deletion candidate; do not delete in the directory phase.
+
+**Later (architecture phase):** see `docs/refactor/04-architecture.md`.
