+++ src/components/sections/features/investment/InvestmentProgress.tsx
