+++ src/components/sections/features/investment/InvestmentTracker.tsx
