+++ src/components/sections/features/investment/InvestmentTrackerContent.tsx
