+++ src/components/sections/features/partners/FriendsInGermany.tsx
