+++ src/components/sections/features/partners/README.md
@@ -1,0 +1,7 @@
+# features/partners/
+
+**Purpose:** Page-section components for partners / friends of the TFFF (NGOs, country friends).
+
+**Here now:** `FriendsInGermany`, `EngagingNGOs`.
+
+**Later (architecture phase):** see `docs/refactor/04-architecture.md`.
