+++ src/components/sections/features/shared/InfoGrid.tsx
